feat: nothing.

+++ npc/csrc/include/common.h
@@ -1,10 +1,16 @@
 #ifndef __COMMON_H__
 #define __COMMON_H__
 
+#include "Vmy_cpu.h"
+#include "Vmy_cpu__Dpi.h"
+#include "svdpi.h"
+#include "verilated.h"
+#include "verilated_vcd_c.h"
 #include <stdio.h>
 #include <assert.h>
+#include <cstdint>
 
-typedef unsigned long long int word_t;
-typedef long long int sword_t;
+typedef uint64_t word_t;
+typedef int64_t sword_t;
 
 #endif
